fixed minor ordering/polishingafter review

+++ contracts/crowdsale/validation/UserCappedCrowdsale.sol
@@ -30,7 +30,7 @@ contract UserCappedCrowdsale is Crowdsale, Ownable {
    * @param _cap Wei limit for individual contribution
    */
   function setGroupCap(address[] _beneficiaries, uint256 _cap) external onlyOwner {
-    for (uint i = 0; i < _beneficiaries.length; i++) {
+    for (uint256 i = 0; i < _beneficiaries.length; i++) {
       caps[_beneficiaries[i]] = _cap;
     }
   }
